refactor: change storage utils
